Week6 (26-30)/Day_29/GenAI Powered E-Commerce Project/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic>=2.7,<3
pydantic-settings==2.5.2
python-dotenv==1.0.1
orjson==3.10.7
openai==1.45.0
chromadb>=1.5.1
pymongo>=4.6.0
pandas>=2.0.0
openpyxl>=3.1.0
httpx==0.27.2
presidio-analyzer==2.2.357
presidio-anonymizer==2.2.357
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.4.0
langsmith>=0.1.0,<0.4.0
